Consolodated button widgets to customWidgets
